Reformat proxy file template
The generated proxy file had poor formatting, this has been cleaned.
The file also included a global constant that could collide with other
proxies. DEFAULT_ODATA_SERVICE_URL was migrated to a class constant.
The require path was updated to be more reasonable to a PHP include
path that is prefixed.